task_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
pydantic
python-multipart
google-generativeai
pydantic-settings